+++ tiledb/sm/filter/filter_pipeline.cc
@@ -255,6 +255,10 @@ Status FilterPipeline::filter_chunks_reverse(
     ThreadPool* const compute_tp,
     const bool unfiltering_all,
     const Config& config) const {
+  if (input.empty()) {
+    return Status::Ok();
+  }
+
   // Precompute the sizes of the final output chunks.
   int64_t chunk_size = 0;
   uint64_t total_size = 0;
@@ -290,7 +294,7 @@ Status FilterPipeline::filter_chunks_reverse(
     if (buffer == nullptr) {
       return LOG_STATUS(Status::FilterError("tdb_malloc() failed"));
     }
-    output->set_contiguous(buffer);
+    RETURN_NOT_OK_ELSE(output->set_contiguous(buffer), tdb_free(buffer));
   }
 
   // Run each chunk through the entire pipeline.